+++ src/responseCrud.js
@@ -1,10 +1,15 @@
-// src/config/responseCrud.js
 const { connect } = require('./config/database');
 
 // Create a new response
 async function createResponse(response) {
   const { db, client } = await connect();
   try {
+    const existingResponse = await db.collection('responses').findOne({ _id: response._id });
+    if (existingResponse) {
+      console.log('Response with this _id already exists');
+      return; 
+    }
+
     const result = await db.collection('responses').insertOne(response);
     console.log(`New response created with the following id: ${result.insertedId}`);
   } finally {
@@ -17,7 +22,9 @@ async function readResponse(responseId) {
   const { db, client } = await connect();
   try {
     const response = await db.collection('responses').findOne({ _id: responseId });
-    console.log(response);
+    if (!response) {
+      console.log('Response not found');
+    }
     return response;
   } finally {
     await client.close();
@@ -28,6 +35,12 @@ async function readResponse(responseId) {
 async function updateResponse(responseId, update) {
   const { db, client } = await connect();
   try {
+    const existingResponse = await db.collection('responses').findOne({ _id: responseId });
+    if (!existingResponse) {
+      console.log('No response found with the given _id');
+      return;
+    }
+
     const result = await db.collection('responses').updateOne({ _id: responseId }, { $set: update });
     console.log(`${result.matchedCount} response(s) matched the filter, updated ${result.modifiedCount} response(s)`);
   } finally {
@@ -39,6 +52,12 @@ async function updateResponse(responseId, update) {
 async function deleteResponse(responseId) {
   const { db, client } = await connect();
   try {
+    const existingResponse = await db.collection('responses').findOne({ _id: responseId });
+    if (!existingResponse) {
+      console.log('No response found with the given _id');
+      return;
+    }
+
     const result = await db.collection('responses').deleteOne({ _id: responseId });
     console.log(`Deleted ${result.deletedCount} response(s)`);
   } finally {
